ariadne/requirements.txt
aiodataloader==0.4.0
anyio==4.3.0
ariadne==0.22
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.3.0
dnspython==2.6.1
graphql-core==3.2.3
h11==0.14.0
idna==3.6
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
setuptools==68.2.2
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
uvicorn==0.28.1
wheel==0.41.2